Add files via upload
I made a number of changes to allow different questions types and multiple js files so that I can add content without always having to go back and fit it into exiting files. The goal is modularity. 
There might be a number of bugs and code that is redundant. I was testing locally for convenience and it required some tweeks.